refactor(tldraw): import TLRichText from tlschema consistently
api-extractor emitted a duplicate `TLRichText_2` because the package's
source imported the type from `@tldraw/editor` while the generated shape
prop types resolved it from `@tldraw/tlschema`. Import from tlschema
everywhere so the API report has a single, named reference.
